% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fortify-mantel.R
\name{fortify_mantel}
\alias{fortify_mantel}
\title{Fortify a mantel's test data}
\usage{
fortify_mantel(spec_df, env_df, env_ctrl = NULL, spec_select = NULL,
  spec_group = NULL, env_group = NULL, env_ctrl_group = NULL,
  pair_test = FALSE, spec_fun = "vegdist", env_fun = "vegdist",
  mantel_fun = "mantel", spec_params = list(method = "bray"),
  env_params = list(method = "euclidean"), env_corr_params = list(type
  = "upper", show_diag = FALSE), get_link_params = list(),
  process = TRUE, ...)
}
\arguments{
\item{spec_df}{a data.frame of species.}

\item{env_df}{a data.frame of environment.}

\item{env_ctrl}{a list of dissimilarity matrices or a data.frame.}

\item{spec_select}{a list of numeric vector to split `spec_df` by columns.}

\item{spec_group}{variable name string to split `spec_df` to a list by rows.}

\item{env_group}{variable name string to split `env_df` to a list by rows.}

\item{env_ctrl_group}{variable name string to split `env_ctrl` data.frame to a list by rows.}

\item{pair_test}{logical, see details.}

\item{spec_fun}{a string of distance function, it is used to handle `gdis` parameter.}

\item{env_fun}{a string of distance function, it is used to handle `mdis` parameter.}

\item{mantel_fun}{a string of mantel's test function.}

\item{spec_params}{a list of extra params passing to `gdist_fun`.}

\item{env_params}{a list of extra params passing to `mdist_fun`.}

\item{env_corr_params}{a list of extra params passing to `ggcorrr::fortify_corr()`.}

\item{get_link_params}{a list of extra params passing to `get_link_data()`.}

\item{process}{logical, whether to print progress information.}

\item{...}{extra params passing to `mantel_fun`.}
}
\value{
a list of `env_df` correalation data, link line data and correalation matrix plot type.
}
\description{
\code{fortify_mantel()}  provides a unified interface for mantel's test, and supports for specifying different
distance function (now support \code{\link[stats]{dist}} and \code{\link[vegan]{vegdist}}) and different mantel's test functions (
(now support \code{\link[vegan]{mantel}}, \code{\link[vegan]{mantel.partial}}.
}
\details{
`spec_df`, `env_df` is splitted to list by row or column. `pair_test` parameter is used to distinguish two different modes.
If TRUE, mantel's test is performed on the i-th element of `spec_df` list  and each element
in the i-th element of `env_df` list. The defaults is FALSE, mantel's test is
performed on the i-th element of `spec_df` list and each elements in `env_df` list.
}
\examples{
library(dplyr)
library(vegan)
data(varespec)
data(varechem)
## four group
df4 <- fortify_mantel(varespec, varechem,
                      spec_select = list(spec01 = 22:25,
                                         spec02 = 1:4,
                                         spec03 = 38:43,
                                         spec04 = 15:20),
                      mantel_fun = "mantel",
                      env_corr_params = list(type = "upper",
                      show_diag = FALSE,
                      corr_test = TRUE,
                      cluster = TRUE))
ggmantel(df4, corr_type = "pie")
## split by columns
spec_df <- varespec \%>\%
             mutate(grp = rep(LETTERS[1:3], 8))

env_df <- varechem \%>\%
            mutate(grp = rep(LETTERS[1:3], 8))

df_grp <- fortify_mantel(spec_df, env_df,
                         spec_group = "grp",
                         env_group = "grp",
                         mantel_fun = "mantel",
                         pair_test = TRUE,
                         env_corr_params = list(type = "upper",
                                                show_diag = FALSE,
                                                corr_test = TRUE,
                                                cluster = TRUE))
ggmantel(df_grp)
}
\seealso{
[vegan::vegdist()], [vegan::mantel()], [vegan::mantel.partial()],
[gghpcc::get_link_data()] and [ggcorrr::fortify_corr()].
}
