% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/link-data.R
\name{get_link_data}
\alias{get_link_data}
\title{Create a link data.frame}
\usage{
get_link_data(mantel_df, corr_df, type = c("upper", "lower"),
  var_grp_name = NULL, var_mat_name = NULL, show_diag = FALSE,
  grp_hjust = NULL, grp_vjust = NULL, mat_hjust = NULL,
  mat_vjust = NULL)
}
\arguments{
\item{mantel_df}{mantel's test data.frame, must be returned by \code{\link[gghpcc]{fortify_mantel}}.}

\item{corr_df}{correalation's test data.frame, must be returned by \code{\link[ggcorrr]{fortify_corr}}.}

\item{type}{correalation matrix plot's type.}

\item{var_grp_name}{the column name of link group variable in `mantel_df`.}

\item{var_mat_name}{the column name of link matrix variable in `mantel_df`.}

\item{show_diag}{logical, if TRUE, correalation matrix plot will don't show diagonal.}

\item{grp_hjust}{a numeric vector to adjust the horizonal position of group points.}

\item{grp_vjust}{a numeric vector to adjust the vertical position of group points.}

\item{mat_hjust}{a numeric vector to adjust the horizonal position of correalation matrixgroup points.}

\item{mat_vjust}{a numeric vector to adjust the vertical position of correalation matrixgroup points.}
}
\value{
a data.frame.
}
\description{
At present, this function is still in the experimental stage. `mantel_df` only
supports the results of \code{\link[gghpccl]{tidy_mantel}} or \code{\link[gghpcc]{tidy_mantel_pair}}, and `corr_df` only
supports the results of  \code{\link[ggcorrr]{fortify_corr}}.
}
\seealso{
[gghpcc::tidy_mantel()], [gghpcc::tidy_mantel_pair()], [ggcorrr::fortify_corr().
}
