% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggmantel.R
\name{ggmantel}
\alias{ggmantel}
\title{Draw mantel test plot}
\usage{
ggmantel(data, corr_type = "square", corr_extra_params = corr_params(),
  corr_backgroud = NA, corr_grid_colour = "grey60",
  stat_breaks = c(-Inf, 0.25, 0.5, Inf), stat_labels = c("< 0.25",
  "0.25 - 0.5", ">= 0.5"), signif_breaks = c(-Inf, 0.001, 0.01, 0.05,
  Inf), signif_labels = c("< 0.001", "0.001 - 0.01", "0.01 - 0.05",
  ">= 0.05"), curvature = NULL, grp_mark = TRUE, grp_mark_shape = 21,
  grp_mark_size = 4, grp_mark_col = "blue", grp_mark_fill = "blue",
  mat_mark = TRUE, mat_mark_shape = 21, mat_mark_size = 4,
  mat_mark_col = "grey60", mat_mark_fill = "grey60",
  grp_label = NULL, grp_label_colour = "black", grp_label_size = 6.5,
  grp_label_family = "", grp_label_fontface = 1,
  grp_label_hspace = 0.2, grp_label_vspace = 0,
  grp_label_hjust = NULL, grp_label_vjust = 0.5,
  scale_fill_col = NULL, scale_colour_col = NULL,
  scale_size_value = c(0.5, 1.5, 4), legend_position = NULL,
  legend_mat_title = NULL, legend_mat_breaks = NULL,
  legend_mat_labels = NULL, x_label_position = NULL,
  y_label_position = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{data}{mantel's test data, now it just support the data returned by `fortify_mantel()`.}

\item{corr_type}{the symbols of correalation matrix plot, see @details .}

\item{corr_extra_params}{extra params pass to correalation matrix plot's layer, see @details .}

\item{corr_backgroud}{background colour of correalation matrix plotting area.}

\item{corr_grid_colour}{correalation matrix plot's grid line colour.}

\item{stat_breaks}{a numeric vector of unique cut points giving the number of intervals into which mantel's
statistic is to be cut, the defaults values is (-Inf, 0.25, 0.5, Inf).}

\item{stat_labels}{labels for the levels of the mantel's statistic cut resulting category.}

\item{signif_breaks}{a numeric vector of unique cut points giving the number of intervals into which mantel's
test p-values is to be cut, the defaults values is (-Inf, 0.001, 0.01, 0.05, Inf).}

\item{signif_labels}{labels for the levels of the mantel's test p-values cut resulting category.}

\item{curvature}{a numeric value giving the amount of curvature. if `type = "upper"`, the default value is -0.1,
else is 0.1.}

\item{grp_mark}{logical. Should add mark symbols on group link points.}

\item{grp_mark_shape}{group mark symbol, the defualt value is 21.}

\item{grp_mark_size}{group mark symbol's size, the defualt value is 4.}

\item{grp_mark_col}{group mark symbol's edge colour, the defualt value is "blue".}

\item{grp_mark_fill}{group mark symbol's fill colour, the defualt value is "blue".}

\item{mat_mark}{logical. Should add mark symbols on  correalation matrix plot's link points.}

\item{mat_mark_shape}{mat mark symbol, the defualt value is 21.}

\item{mat_mark_size}{mat mark symbol's size, the defualt value is 4.}

\item{mat_mark_col}{mat mark symbol's edge colour, the defualt value is "grey60".}

\item{mat_mark_fill}{mat mark symbol's fill colour, the defualt value is "grey60".}

\item{grp_label}{character vecter, group link labels.}

\item{grp_label_colour}{group link labels' colour.}

\item{grp_label_size}{group link labels' size.}

\item{grp_label_family}{group link labels' text family.}

\item{grp_label_fontface}{group link labels' text font face.}

\item{grp_label_hspace}{the horizontal space between group link points and labels.}

\item{grp_label_vspace}{the vertical space between group link points and labels.}

\item{grp_label_hjust}{group link labels' horizontal justification.}

\item{grp_label_vjust}{group link labels' vertical justification.}

\item{scale_fill_col}{a colour string vector, the default value is `ggcorrr:::.default_colors`.}

\item{scale_colour_col}{a colour string vector.}

\item{scale_size_value}{a numeric vector.}

\item{legend_position}{the position of legends ("none", "left", "right", "bottom", "top", or two-element numeric vector).}

\item{legend_mat_title}{title of legend colourbar.}

\item{legend_mat_breaks}{a numeric vector of positions.}

\item{legend_mat_labels}{a character vector giving labels (must be same length as legend_mat_breaks).}

\item{x_label_position}{the position of x axis. "top" or "bottom".}

\item{y_label_position}{the position of y axis. "left" or "right".}

\item{xlim}{a numeric vector of length two providing limits of the x axis scale.}

\item{ylim}{a numeric vector of length two providing limits of the y axis scale.}
}
\description{
`ggmantel()` is highly dependent on `ggcorrr` packages, and it provides a simple,
flexible and easy-to-use function interface to draw mantel's test graphs. I strongly recommend
that people who are not familiar with `ggplot2`'s grammar use `ggmantel()` directly to draw mantel's
test plot. And at the same time, in order to maintain flexibility and customization, people
who are very familiar with `ggplot2`'s grammar can also just use `fortify_mantel()` to handle data. .
}
\details{
`corr_type` supports the following nine styles and combinations.
\itemize{
    \item{"square", @seealso [ggcorrr::geom_square()].}
    \item{"colour", @seealso [ggcorrr::geom_colour].}
    \item{"circle", @seealso [ggcorrr::geom_circle2()].}
    \item{"ellipse", @seealso [ggcorrr::geom_ellipse2()].}
    \item{"pie", @seealso [ggcorrr::geom_pie2()].}
    \item{"text", @seealso [ggcorrr::geom_num()].}
    \item{"confbox", @seealso [ggcorrr::geom_confbox()].}
    \item{"mark", @seealso [ggcorrr::geom_mark()].}
    \item{"shade", @seealso [ggcorrr::geom_shade()].}
    \item{"cross", @seealso [ggcorrr::geom_cross()].}
}

Using "+" symbols to combine multiple layers, overlapping from left to right.

`corr_extra_params` must be list. If there are multiple layers, we can set additional params for each layer with a list.
When a layer does not need additional parameters, an empty list needs to be added.
}
\examples{
library(vegan)
data(varespec)
data(varechem)
## four group
df4 <- fortify_mantel(varespec, varechem,
                      spec_select = list(spec01 = 22:25,
                                         spec02 = 1:4,
                                         spec03 = 38:43,
                                         spec04 = 15:20),
                      mantel_fun = "mantel",
                      env_corr_params = list(type = "upper",
                      show_diag = FALSE,
                      corr_test = TRUE,
                      cluster = TRUE))
ggmantel(df4)
}
